refactor: avancée sur l'aspect accéssible du site web
