+++ css/homepage/header.css
@@ -41,7 +41,7 @@ header h1 {
     }
 
     header nav ul {
-        top: 30px;
+        top: 22px;
         justify-content: center;
     }
