+++ css/homepage/photographers-section.css
@@ -18,8 +18,8 @@
 }
 
 #photographers img {
-    width: 11rem;
-    height: 11rem;
+    width: 12.5rem;
+    height: 12.5rem;
     object-fit: cover;
     background-color: rgb(89, 206, 38);
     margin: 0 auto;
@@ -36,7 +36,7 @@
 
 .name {
     color: #D3573C;
-    font-size: 1.5rem;
+    font-size: 1.7rem;
     margin-bottom: 0.7rem;
 }
