+++ index.html
@@ -13,8 +13,8 @@
 
 <body>
     <!-- BANNER -->
-    <header role="banner">
-        <a href="index.html"><img src="https://res.cloudinary.com/dlpyn3wxf/image/upload/v1614078430/logo_mig5no.png"
+    <header role="banner" aria-hidden='false'>
+        <a href="index.html" title="HomePage"><img src="https://res.cloudinary.com/dlpyn3wxf/image/upload/v1614078430/logo_mig5no.png"
                 alt="FishEye Home Page"></a>
         <!-- NAV -->
         <nav class="filters" aria-label="photographer categories" role="navigation">
@@ -32,7 +32,7 @@
         <h1>Nos Photographes</h1>
     </header>
     <!-- BUTTON 'PASSER AU CONTENU' -->
-    <a id="main-photographers_link" href="#">Passer au contenu</a>
+    <a id="main-photographers_link" href="#photographers">Passer au contenu</a>
     <main>
         <!-- PHOTOGRAPHERS -->
         <section id="photographers" aria-label="photographers">
