+++ js/main.js
@@ -9,6 +9,7 @@ import {} from './home/filter.js';
 import PhotographerProfil from './photographers-page/PhotographerProfil.js';
 import DropDown from './photographers-page/DropDown.js';
 import Works from './photographers-page/Works.js';
+import Likes from './photographers-page/likes.js';
 import Form from './photographers-page/form.js';
 
 // PHOTOGRAPHER PROFIL HEADER
@@ -16,6 +17,9 @@ import Form from './photographers-page/form.js';
 
 // PHOTOGRAPHER GALLERY & LIKES
 (new Works()).photographersWorks();
+
+// LIKES & BOX
+(new Likes()).likes();
 
 // DROPDOWN MENU
 (new DropDown()).dropDown();
