+++ js/photographers-page/Works.js
@@ -14,7 +14,7 @@ export default class Works {
         let totalLike = 0;
         let currentMedia = [];
         let currentMediaName = [];
-        let currentLightboxIndex = 0;
+        let currentLightboxIndex = null;
 
         media.forEach(element => {
             if (id == element.photographerId) {
@@ -30,8 +30,7 @@ export default class Works {
                     <span class="ph-work-like">
                         <a class="like-counter">${element.likes}</a>
                     </span>
-                    <button class="btn-like" type="button">
-                        <i class="fas fa-heart btn" aria-label='likes' role="button" data-value="${element.likes}"></i>
+                    <button class="btn-like" type="button"><i class="fas fa-heart heart-btn" aria-label='likes' role="button" data-value="${element.likes}"></i>
                     </button>
                     </div>
                 </div>
@@ -48,52 +47,5 @@ export default class Works {
                 (new Lightbox()).lightboxKeyboard(currentMedia, currentMediaName, currentLightboxIndex);
             }
         })
-
-        /*
-        function boxLikesAndPrice(data, id, totalLike) {
-            const photographers = data.photographers;
-
-            photographers.forEach(element => {
-                if (id == element.id) {
-                    let box = document.getElementById('box');
-                    let boxTemplate = `
-                    <span id="total-likes">${totalLike}  <i class="fas fa-heart" aria-label='likes'></i>
-                    </span>
-                    <span>${element.price} €/ jour</span>
-                    `
-                    box.innerHTML = boxTemplate;
-                }
-            })
-        }
-        */
-
-        /*
-        function likes(totalLike) {
-            const div = document.getElementById('ph-works');
-            boxLikesAndPrice(data, id, totalLike);
-
-            div.addEventListener('click', function (e) {
-                let buttonsLike = document.querySelectorAll('.btn');
-
-                buttonsLike.forEach(function (btn) {
-                    let totalLikes = document.getElementById('total-likes');
-                    let initElem = e.target;
-                    let likeValue = initElem.dataset.value;
-                    let likesCounter = initElem.parentNode.parentNode.firstElementChild;
-                    let classValue = btn.classList.value.split(' ');
-                    let classTarget = e.target.classList.value.split(' ');
-                    let intersection = classTarget.filter(x => classValue.includes(x));
-
-                    if (intersection) {
-                        likeValue++;
-                        totalLike++;
-                        likesCounter.innerHTML = likeValue;
-                        totalLikes.innerHTML = totalLike + ` <i class="fas fa-heart" aria-label='likes'></i>`;
-                    }
-                })
-            })
-        }
-        likes(totalLike);
-        */
     }
 }
